requirements.txt
accelerate==0.26.1
certifi==2024.7.4
charset-normalizer==3.3.2
diffusers==0.32.2
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.24.3
idna==3.7
importlib-metadata==7.0.1
Jinja2==3.1.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
packaging==23.2
peft==0.9.0
pillow==10.3.0
psutil==5.9.7
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.2
safetensors==0.4.1
sympy==1.12
tokenizers==0.21.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.3
transformers==4.48.0
triton==3.0.0
typing_extensions==4.9.0
urllib3==2.2.2
zipp==3.19.1
sentencepiece==0.2.0
bitsandbytes==0.45.4
GPUtil==1.4.0
scipy==1.11.2
protobuf==6.30.2
